Add Minimise and Maximise Button functionality
The minimise button changes the toolbar into a small form factor so that it blocks less of the screen. In the mimised state, the only remaining button on the toolbar is a new Maximise button. Clicking this will restore the toolbar to the full format. The maximise color also changes color similarly to the Activate Drawing Mode button so that users can immediately see whether they are in drawing mode